@@ -1,20 +1,20 @@
 {
-  "name": "phonetworks/pho-lib-graph",
-  "type": "library",
-  "description": "A general purpose graph library in PHP",
-   "keywords": [
+    "name": "phonetworks/pho-lib-graph",
+    "type": "library",
+    "description": "A general purpose graph library in PHP",
+    "keywords": [
         "graph",
         "network",
         "node",
         "vertex",
         "edge",
-	    "predicate"
+        "predicate"
     ],
-  "homepage": "https://phonetworks.org",
-  "license": "MIT",
-  "prefer-stable": true,
-  "minimum-stability": "dev",
-  "authors": [
+    "homepage": "https://phonetworks.org",
+    "license": "MIT",
+    "prefer-stable": true,
+    "minimum-stability": "dev",
+    "authors": [
         {
             "name": "Emre Sokullu",
             "email": "emre@phonetworks.org",
@@ -22,27 +22,27 @@
         }
     ],
     "require": {
-        "php":">=7.1",
-        "psr/log": "^3.0",
-        "phonetworks/pho-lib-dht": "^1.0"
+        "php": ">=8.2",
+        "psr/log": "*",
+        "ramsey/uuid": "^4.7.6"
     },
     "require-dev": {
-        "phpunit/phpunit": "^5.7",
-	    "phpdocumentor/phpdocumentor": "*" ,
+        "phpunit/phpunit": "*",
+        "phpdocumentor/phpdocumentor": "^3.0",
         "psy/psysh": "*",
-        "fzaninotto/faker": "^1.6",
-        "mockery/mockery": "dev-master",
-	    "squizlabs/php_codesniffer": "3.*",
+        "fzaninotto/faker": "dev-master",
+        "mockery/mockery": "^1.0",
+        "squizlabs/php_codesniffer": "3.*",
         "roave/security-advisories": "dev-master"
-        },
+    },
     "autoload": {
-	     "psr-4": {
-		       "Pho\\Lib\\Graph\\": "src/Pho/Lib/Graph/"
-	     }
+        "psr-4": {
+            "Pho\\Lib\\Graph\\": "src/Pho/Lib/Graph/"
+        }
     },
     "autoload-dev": {
-	     "psr-4": {
-		       "Pho\\Lib\\Graph\\": "tests/Pho/Lib/Graph/"
-	     }
+        "psr-4": {
+            "Pho\\Lib\\Graph\\": "tests/Pho/Lib/Graph/"
+        }
     }
 }
