@@ -11,23 +11,24 @@
 
 namespace Pho\Lib\Graph;
 
-class LoggerTest extends TestCase 
-{
-    public function setUp() {
-        parent::setUp();
-        Logger::setVerbosity(2);
-    }
+class LoggerTest extends TestCase {
 
-    public function tearDown() {
-        Logger::setVerbosity(0);
-        parent::tearDown();
-    }
+  public function setUp(): void {
+    parent::setUp();
+    Logger::setVerbosity(2);
+  }
 
-    public function testLog() {
-        ob_start();
-        $node = new Node($this->graph);
-        $output = ob_get_clean();
-        $expected = "A node with id \"";
-        $this->assertEquals($expected, substr($output,0,strlen($expected)));
-    }
+  public function tearDown(): void {
+    Logger::setVerbosity(0);
+    parent::tearDown();
+  }
+
+  public function testLog() {
+    ob_start();
+    $node = new Node($this->graph);
+    $output = ob_get_clean();
+    $expected = "A node with id \"";
+    $this->assertEquals($expected, substr($output, 0, strlen($expected)));
+  }
+
 }
